+++ src/Network/Network/AzureFirewall/NewAzureFirewallCommand.cs
@@ -193,15 +193,26 @@ namespace Microsoft.Azure.Commands.Network
             HelpMessage = "A list of availability zones denoting where the firewall needs to come from.")]
         public string[] Zone { get; set; }
 
+        [Alias("Sku")]
         [Parameter(
-                Mandatory = false,
-                ValueFromPipelineByPropertyName = true,
-                HelpMessage = "The sku type for firewall")]
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The sku name for firewall")]
         [ValidateSet(
                 MNM.AzureFirewallSkuName.AZFWHub,
                 MNM.AzureFirewallSkuName.AZFWVNet,
                 IgnoreCase = false)]
-        public string Sku { get; set; }
+        public string SkuName { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The sku tier for firewall")]
+        [ValidateSet(
+                MNM.AzureFirewallSkuTier.Standard,
+                MNM.AzureFirewallSkuTier.Premium,
+                IgnoreCase = false)]
+        public string SkuTier { get; set; }
 
         [Parameter(
                 Mandatory = false,
@@ -262,7 +273,11 @@ namespace Microsoft.Azure.Commands.Network
         private PSAzureFirewall CreateAzureFirewall()
         {
             var firewall = new PSAzureFirewall();
-            if (Sku == MNM.AzureFirewallSkuName.AZFWHub)
+            var sku = new PSAzureFirewallSku();
+            sku.Name = !string.IsNullOrEmpty(this.SkuName) ? this.SkuName: MNM.AzureFirewallSkuName.AZFWVNet;
+            sku.Tier = !string.IsNullOrEmpty(this.SkuTier) ? this.SkuTier : MNM.AzureFirewallSkuTier.Standard;
+
+            if (this.SkuName == MNM.AzureFirewallSkuName.AZFWHub)
             {
 
                 if (VirtualHubId != null && this.Location != null)
@@ -281,10 +296,6 @@ namespace Microsoft.Azure.Commands.Network
                     throw new ArgumentException("The list of public Ip addresses cannot be provided during the firewall creation");
                 }
 
-                var sku = new PSAzureFirewallSku();
-                sku.Name = MNM.AzureFirewallSkuName.AZFWHub;
-                sku.Tier = MNM.AzureFirewallSkuTier.Standard;
-
                 firewall = new PSAzureFirewall()
                 {
                     Name = this.Name,
@@ -298,9 +309,6 @@ namespace Microsoft.Azure.Commands.Network
             }
             else
             {
-                var sku = new PSAzureFirewallSku();
-                sku.Name = MNM.AzureFirewallSkuName.AZFWVNet;
-                sku.Tier = MNM.AzureFirewallSkuTier.Standard;
                 firewall = new PSAzureFirewall()
                 {
                     Name = this.Name,