+++ src/Network/Network/ChangeLog.md
@@ -20,7 +20,14 @@
 
 ## Upcoming Release
 * [Breaking Change] Removed parameter `HostedSubnet` and add `Subnet` instead
-
+* Added new cmdlets for Virtual Router Peer Routes
+    - `Get-AzVirtualRouterPeerLearnedRoute`
+    - `Get-AzVirtualRouterPeerAdvertisedRoute`
+* Updated New-AzFirewall cmdlet:
+    - Added parameter -SkuTier
+    - Added parameter -SkuName and made Sku as Alias for this
+    - Removed parameter -Sku
+    
 ## Version 3.5.0
 * Added Office365 Policy to VPNSite Resource
     - `New-AzO365PolicyProperty`